cloud_poster/requirements.txt
fastapi
motor
python-dotenv
uvicorn
pydantic
python-multipart